setup router, crud and db setup
