+++ README.md
@@ -1,2 +1,7 @@
 # patient-appointment-backend
 Repository for the Patient Appointment Network Data Application (PANDA) backend service.
+
+TODO:
+- Deployment instructions
+- Tech choices and considerations
+- Development thought process
