+++ app/api/main.py
@@ -1,0 +1,6 @@
+from fastapi import APIRouter
+
+from app.api.routes import patients
+
+api_router = APIRouter()
+api_router.include_router(patients.router)
