+++ todo.txt
@@ -1,0 +1,12 @@
+Impl Routes
+Impl CRUD
+Environment variables
+NHS Number checksum
+API Docs
+Tests
+README
+
+Done:
+Routes setup
+CRUD setup
+core/db setup
